Revert "Full Worlddb Int Cleanup, Base of this fix By TrinityCore"

+++ src/server/game/Events/GameEventMgr.cpp
@@ -222,7 +222,7 @@ void GameEventMgr::LoadFromDB()
         {
             Field* fields = result->Fetch();
 
-            uint16 event_id = fields[0].GetUInt8();
+            uint16 event_id = fields[0].GetUInt16();
             if (event_id == 0)
             {
                 sLog->outErrorDb("`game_event` game event entry 0 is reserved and can't be used.");
@@ -234,8 +234,8 @@ void GameEventMgr::LoadFromDB()
             pGameEvent.start        = time_t(starttime);
             uint64 endtime          = fields[2].GetUInt64();
             pGameEvent.end          = time_t(endtime);
-            pGameEvent.occurence    = fields[3].GetUInt64();
-            pGameEvent.length       = fields[4].GetUInt64();
+            pGameEvent.occurence    = fields[3].GetUInt32();
+            pGameEvent.length       = fields[4].GetUInt32();
             pGameEvent.holiday_id   = HolidayIds(fields[5].GetUInt32());
 
             pGameEvent.state        = (GameEventState)(fields[7].GetUInt8());
@@ -317,7 +317,6 @@ void GameEventMgr::LoadFromDB()
     {
         uint32 oldMSTime = getMSTime();
 
-        //                                                   0             1
         QueryResult result = WorldDatabase.Query("SELECT eventEntry, prerequisite_event FROM game_event_prerequisite");
         if (!result)
         {
@@ -331,7 +330,7 @@ void GameEventMgr::LoadFromDB()
             {
                 Field* fields = result->Fetch();
 
-                uint16 event_id = fields[0].GetUInt8();
+                uint16 event_id = fields[0].GetUInt16();
 
                 if (event_id >= mGameEvent.size())
                 {
@@ -341,7 +340,7 @@ void GameEventMgr::LoadFromDB()
 
                 if (mGameEvent[event_id].state != GAMEEVENT_NORMAL && mGameEvent[event_id].state != GAMEEVENT_INTERNAL)
                 {
-                    uint16 prerequisite_event = fields[1].GetUInt32();
+                    uint16 prerequisite_event = fields[1].GetUInt16();
                     if (prerequisite_event >= mGameEvent.size())
                     {
                         sLog->outErrorDb("`game_event_prerequisite` game event prerequisite id (%i) is out of range compared to max event id in `game_event`", prerequisite_event);
@@ -385,7 +384,7 @@ void GameEventMgr::LoadFromDB()
                 Field* fields = result->Fetch();
 
                 uint32 guid    = fields[0].GetUInt32();
-                int16 event_id = fields[1].GetInt8();
+                int16 event_id = fields[1].GetInt16();
 
                 int32 internal_event_id = mGameEvent.size() + event_id - 1;
 
@@ -411,7 +410,7 @@ void GameEventMgr::LoadFromDB()
     {
         uint32 oldMSTime = getMSTime();
 
-        //                                                      0                1
+        //                                                      1                2
         QueryResult result = WorldDatabase.Query("SELECT gameobject.guid, game_event_gameobject.eventEntry FROM gameobject"
                                                  " JOIN game_event_gameobject ON gameobject.guid=game_event_gameobject.guid");
 
@@ -428,7 +427,7 @@ void GameEventMgr::LoadFromDB()
                 Field* fields = result->Fetch();
 
                 uint32 guid    = fields[0].GetUInt32();
-                int16 event_id = fields[1].GetInt8();
+                int16 event_id = fields[1].GetInt16();
 
                 int32 internal_event_id = mGameEvent.size() + event_id - 1;
 
@@ -528,7 +527,7 @@ void GameEventMgr::LoadFromDB()
 
                 uint32 id       = fields[0].GetUInt32();
                 uint32 quest    = fields[1].GetUInt32();
-                uint16 event_id = fields[2].GetUInt8();
+                uint16 event_id = fields[2].GetUInt16();
 
                 if (event_id >= mGameEventCreatureQuests.size())
                 {
@@ -569,7 +568,7 @@ void GameEventMgr::LoadFromDB()
 
                 uint32 id       = fields[0].GetUInt32();
                 uint32 quest    = fields[1].GetUInt32();
-                uint16 event_id = fields[2].GetUInt8();
+                uint16 event_id = fields[2].GetUInt16();
 
                 if (event_id >= mGameEventGameObjectQuests.size())
                 {
@@ -593,7 +592,7 @@ void GameEventMgr::LoadFromDB()
     {
         uint32 oldMSTime = getMSTime();
 
-        //                                                 0       1         2             3
+        //                                                 0       1         2           3
         QueryResult result = WorldDatabase.Query("SELECT quest, eventEntry, condition_id, num FROM game_event_quest_condition");
 
         if (!result)
@@ -609,7 +608,7 @@ void GameEventMgr::LoadFromDB()
                 Field* fields = result->Fetch();
 
                 uint32 quest     = fields[0].GetUInt32();
-                uint16 event_id  = fields[1].GetUInt8();
+                uint16 event_id  = fields[1].GetUInt16();
                 uint32 condition = fields[2].GetUInt32();
                 float num       = fields[3].GetFloat();
 
@@ -651,7 +650,7 @@ void GameEventMgr::LoadFromDB()
             {
                 Field* fields = result->Fetch();
 
-                uint16 event_id  = fields[0].GetUInt8();
+                uint16 event_id  = fields[0].GetUInt16();
                 uint32 condition = fields[1].GetUInt32();
 
                 if (event_id >= mGameEvent.size())
@@ -662,8 +661,8 @@ void GameEventMgr::LoadFromDB()
 
                 mGameEvent[event_id].conditions[condition].reqNum = fields[2].GetFloat();
                 mGameEvent[event_id].conditions[condition].done = 0;
-                mGameEvent[event_id].conditions[condition].max_world_state = fields[3].GetUInt16();
-                mGameEvent[event_id].conditions[condition].done_world_state = fields[4].GetUInt16();
+                mGameEvent[event_id].conditions[condition].max_world_state = fields[3].GetUInt32();
+                mGameEvent[event_id].conditions[condition].done_world_state = fields[4].GetUInt32();
 
                 ++count;
             }
@@ -742,7 +741,7 @@ void GameEventMgr::LoadFromDB()
                 Field* fields = result->Fetch();
 
                 uint32 guid     = fields[0].GetUInt32();
-                uint16 event_id = fields[1].GetUInt8();
+                uint16 event_id = fields[1].GetUInt16();
                 uint32 npcflag  = fields[2].GetUInt32();
 
                 if (event_id >= mGameEvent.size())
@@ -766,7 +765,7 @@ void GameEventMgr::LoadFromDB()
     {
         uint32 oldMSTime = getMSTime();
 
-        //                                                  0          1
+        //                                               0     1
         QueryResult result = WorldDatabase.Query("SELECT questId, eventEntry FROM game_event_seasonal_questrelation");
 
         if (!result)
@@ -782,7 +781,7 @@ void GameEventMgr::LoadFromDB()
                 Field* fields = result->Fetch();
 
                 uint32 questId  = fields[0].GetUInt32();
-                uint32 eventEntry = fields[1].GetUInt32(); // TODO: Change to uint8
+                uint16 eventEntry = fields[1].GetUInt16();
 
                 if (!sObjectMgr->GetQuestTemplate(questId))
                 {
@@ -825,7 +824,7 @@ void GameEventMgr::LoadFromDB()
             {
                 Field* fields = result->Fetch();
 
-                int8 event_id  = fields[0].GetInt8();
+                uint16 event_id  = fields[0].GetUInt16();
 
                 if (event_id >= mGameEventVendors.size())
                 {
@@ -837,7 +836,7 @@ void GameEventMgr::LoadFromDB()
                 NPCVendorEntry newEntry;
                 uint32 guid = fields[1].GetUInt32();
                 newEntry.item = fields[2].GetUInt32();
-                newEntry.maxcount = fields[3].GetUInt32();
+                newEntry.maxcount = fields[3].GetInt32();
                 newEntry.incrtime = fields[4].GetUInt32();
                 newEntry.ExtendedCost = fields[5].GetUInt32();
                 // get the event npc flag for checking if the npc will be vendor during the event or not
@@ -891,7 +890,7 @@ void GameEventMgr::LoadFromDB()
             {
                 Field* fields = result->Fetch();
 
-                uint16 event_id = fields[0].GetUInt8();
+                uint16 event_id = fields[0].GetUInt16();
 
                 if (event_id >= mGameEvent.size())
                 {
@@ -914,7 +913,7 @@ void GameEventMgr::LoadFromDB()
     {
         uint32 oldMSTime = getMSTime();
 
-        //                                                               0                         1
+        //                                                       1                       2
         QueryResult result = WorldDatabase.Query("SELECT pool_template.entry, game_event_pool.eventEntry FROM pool_template"
                                                  " JOIN game_event_pool ON pool_template.entry = game_event_pool.pool_entry");
 
@@ -931,7 +930,7 @@ void GameEventMgr::LoadFromDB()
                 Field* fields = result->Fetch();
 
                 uint32 entry   = fields[0].GetUInt32();
-                int16 event_id = fields[1].GetInt8();
+                int16 event_id = fields[1].GetInt16();
 
                 int32 internal_event_id = mGameEvent.size() + event_id - 1;
 
@@ -1022,7 +1021,7 @@ void GameEventMgr::StartArenaSeason()
     }
 
     Field* fields = result->Fetch();
-    uint16 eventId = fields[0].GetUInt8();
+    uint16 eventId = fields[0].GetUInt16();
 
     if (eventId >= mGameEvent.size())
     {
